Extract send_decision_email test to new file
